pyproject.toml
[project]
name = "samatcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "decord>=0.6.0",
    "einops>=0.8.1",
    "h5py>=3.13.0",
    "hf-transfer>=0.1.9",
    "huggingface-hub>=0.31.1",
    "hydra-core>=1.3.2",
    "iopath>=0.1.10",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "nvitop>=1.5.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.6.1",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "wandb>=0.19.11",
    "yacs>=0.1.8",
]
